+++ Config.lua
@@ -870,6 +870,7 @@ DF.PartyDefaults = {
     bossDebuffsContainerOverlayGradientDir = 0,
     bossDebuffsContainerOverlayAlpha = 1.0,
     bossDebuffsContainerOverlayFrameLevel = 6,
+    bossDebuffsContainerOverlayStrata = "MEDIUM",
 
     -- Buff settings
     buffAlpha = 1,
@@ -2186,6 +2187,7 @@ DF.RaidDefaults = {
     bossDebuffsContainerOverlayGradientDir = 0,
     bossDebuffsContainerOverlayAlpha = 1.0,
     bossDebuffsContainerOverlayFrameLevel = 6,
+    bossDebuffsContainerOverlayStrata = "MEDIUM",
 
     -- Buff settings
     buffAlpha = 1,